Console Application Started & Fixes
-Fixed/Changed the schedule's vector from vector<Task> to vector<Task *>
(pointers to Task objects) to conform to c++ practices and to allow for
polymorphism (as a vector of task objects, different subclasses of Task
were all using the base class methods rather than the overridden ones).
-Fixed all methods involving the schedule's vector to reflect the
previous change. Made sure that the now dynamically allocated task
objects were deleted from memory whenever being removed from the vector
by these methods (to prevent memory leaks).
-Added a crude console system to add different types of tasks to the
schedule and to fill in the fields of the task objects, as well as to
print the contents of the schedule's vector. Only job/payment tasks
being added has been finished so far.
-Changed the main method to more appropriately reflect the intended MVC
architecture. It now only initializes a model and a console_controller.
-Transferred all the code for displaying the current time to the
console_controller class from the main method. Wrote the beginnings of a
basic console user interface for testing the model until a view is
written.