Finished Schedule sort, Removal/Updating of Tasks that have passed
-Finished the Schedule sortTasks() method so that it sorts by seconds as
well (std::sort was used in place of counting sort for seconds as
counting sort would require large amounts of memory for even small
input).
-Made the database interfacing methods (create, update, delete, read)
for all database wrapper classes public instead of private
-Added an updateSchedule() method to the Schedule class to control the
finding of Tasks that have already passed in the vector of Tasks and
removing them, updating the time in the corresponding database entry to
reflect the next occurrence of the Task iff repeating, or removing the
entry otherwise.