Bills DB methods progress
-Began integrating the database connectivity into the schedule/model's
functions
-Laid groundwork for integrating assignment db methods easily when they
are merged to the master (will be wrapped up in methods in the
dbmanipulator class)
-Fixed many bugs involving passing task subclasses to the dbmanipulator
being sliced, incorrect SQL statements, missing selecting of schema, a
couple peculiar problems that resulted in memory access violations
-Began fixing problems where prep->setString(...) causes memory access
errors for an unknown reason, tried switching to setBlob (seen in
updatePayment), as I read it suggested in some MySQL C++ connector
literature.
-Began implementing adding tasks to database from the program, and
deleting (still bugs to fix)
-Implemented updating the time of repeating tasks to the next
occurrence, and that new time updating in the database
-Made it so all tasks have a rowID field (now in the task class) and
wrote constructors to take in this rowID (so tasks read from the
database can have the rowID put in, without having a setRowID method
which could result in unintended behavior if misused).
-Probably lots of other small things that I've forgotten.